requirements.txt
flask
flask-cors
pillow
numpy
tflite-runtime
gunicorn
